Add check for task queue size to task creation

+++ pkg/model/config.go
@@ -7,6 +7,6 @@ type Config struct {
 
 // ManagerInfo : config fo the manager section
 type ManagerInfo struct {
-	ExecutionQueueSize int `toml:"execution_queue_size"`
-	TaskQueueSize      int `toml:"task_queue_size"`
+	ExecutionQueueSize int64 `toml:"execution_queue_size"`
+	TaskQueueSize      int64 `toml:"task_queue_size"`
 }
